refactor language handling
Because Qt translations need to be initialized before any widgets are created,
and because the Qt language needs to match the gettext language in order for
translated shortcuts to work, per-profile language settings aren't possible.
So instead of storing the language in the profile, we use
pm.meta['defaultLang'] for all profiles and initialize language handling in
__init__.py

The language selection in the preferences has been removed, because in a
school setting a student fiddling with the language could potentially cause
other students to be unable to navigate the UI. Instead, Anki will accept
-l/--lang passed on the command line to override the original language chosen
at first program startup.