don't decode file argument as unicode; tell user to file>import if broken
When the file to import has non-latin characters in its path, we're not
handling the provided filename correctly. I tried decoding with
sys.getfilesystemencoding() with no luck, so this needs more looking into. In
the mean time, if we can't open the path name, tell the user to File>Import
the file instead.