Do not consider as unused images exclusively referenced from templates
When removing unused media, Anki only pays attention to whether the
files in the media directory are referenced from a note. However, it
may be the case that a file is solely referenced from a template. For
example, if we wanted to ask the user to locate a state on a US map,
our Front Template could look like this:

  <img src="empty-us-map.jpg">
  <hr>
  {{StateName}}

And the Back Template could be:

  {{StateLocation}}
  <hr>
  {{StateName}}

The image empty-us-map.jpg, if not referenced from any note, is
reported by Anki as unused. Modify MediaManager.allMedia() so that it
also returns those images referenced (HTML <img> tags) from templates.