shared random number between question and answer sides
It's unfeasible (and pretty useless) to generate cards for numbers individually. Having a shared random value between the question and the answer makes it possible to make cards like "numbers between 1 and 10" or for "10, 20, ..., 100"; javascript can generate both the number and the answer in the foreign language. Consider for example this:

    <script>
      // get a number and the number of zeroes between the set limits
      var number = Math.floor({{Rand}} * ({{Max}} - {{Min}} + 1) ) + {{Min}};
      var zeroes = Math.floor({{Rand2}} * ({{Zeroes Max}} - {{Zeroes Min}} + 1)) + {{Min}};

      // compute the final number
      number = (number+{{Increment By}})*{{Multiplier}}*Math.pow(10,zeroes-1);

      document.getElementById("question").innerHTML = number;
    </script>