Enter/return on Preferences should close window
Explicitly sets the "default" button on the Preferences window to ensure
that a user who hits enter/return gets the right behavior.

QDialogButtonBox arranges its buttons based on the platform the software
is running, and typically maps the enter/return key to the "Ok" button.
Because the Preferences window in Anki does not technically have an "Ok"
button, and instead uses a "Close" button, the enter/return key could
activate the "Help" button instead of the "Close" on platforms that have
"Help" as the left-most button.