Indented text importer
Sometimes it's just easier to compose cards in a standard text
editor. This import format uses a simple indenting scheme that allows
you to maintain line breaks for improved readability. Fields begin at
the start of a line and every indented line that follows is the field
value. Comment lines are supported, blank lines are
ignored. Optionally you can specify the tags and fields using the same
simple indented format.

See the comment at the beginning of the file for more detailed
documentation and examples.

Signed-off-by: John Dennis <jdennis@sharpeye.com>