Allow add-on authors to more easily extend Anki's web views
Extends all pertinent web views with runFilter hooks that enable
add-on authors to easily prepend and append HTML to the body of the
view, or add their own CSS or JS to the head element.

Also implements two runFilter hooks specific to the stats areas
of the Overview and DeckBrowser, granting more control over these
commonly modified parts of Anki's UI.

These changes should help alleviate some of the conflicts between
add-ons that target the same HTML views, and also do away with
some of the hacks that are currently in use to achieve similar
goals.