Making rendering (and requirements) quicker
This code from add-on 115825506. No bug report were sent (only 8
downloads)

On a note type with a hundred card type and field, this improve the
computation of req from 7 minutes and a half to less than 3 minutes.

For badly formatted card, it may change the error message. Otherwise,
it should return the same card. One exception beeing if a tag occur in
the value of a field, in this case this won't be replaced
recursively.