Recompute req only when it may be required.
Currently, anki.models.save recomputer the requirements of the model
saved. (Assuming a model is given as argument).

This computation may be long when there are many card type and
fields. And furthermore, this computation is often useless.  As an
example, on my computer, it takes 4 seconds for the model I use the most.

Let me give you one important example. Each time a new note is
created, the model is changed so that the new deck id of the model is
the current deck id. It seems quite clear that there is absolutely no
reason here to recompute the requirements. It means that between each
note I create, I should wait 4 seconds uselessly.

Thus, I added an optional parameter to ```anki.models.save```,
```recomputeReq```, which simply state whether the requirements should
be recomputed or not. By default, its value is True, thus the default
value ensure that the method act as before. Hence, an add-on using
this method won't see any change.

I also reviewed every occurrences of ```mm.save```,
```models.save``` in which a model is passed as argument, and I wrote
explicitly whether requirements should be recomputed or not.

A similar change was done to ```aqt.models.saveModel```,  where I
have chosen that the requirements are recomputed when the note type
window is closed (instead of recomputing it each time we decide to see
another card type)