Cancel in "adding note type" debugging
There is what I believe to be a bug in anki.
=====

* open the note type manager
* click on add
* Select anything
* click Ok
* click cancel

The expected result would be that the new note type is not added. The
obtained result is that the note type have the default
name. I.e. either old name followed by " copy". Or a name with five
random symbols at the end.

This commit simply ensure that, when the "cancel" button is clicked,
no note type is added.

Correction
=======
There are many way to do it. I did what I think to be the most elegant
way to do it.

The problem I faced that the default name, presented to the user, was
found in the model. However, the standard models does not necessarily
exists yet. And the methods which creates the standard models also add
them to the collection. Since I don't want to change those methods, my
solution was to find the name somewhere else.

That is why the parameter models in aqt.models.AddModel now contains an
list with three values. The third one being the name shown to the
user (unless this name is not unique).

This also means that I did move the request for a name to AddModel,
instead of having it in "onAdd". This ensure that AddModel does the
entire work of adding a model. onAdd simply ass AddModel to do all the
work, and potentially refresh the list.

In order to give the user a name which is unique without having to
create a model, I did split the method ensureNameUnique. It's main
body, i.e. computing a string which is not a name, is now in
ensureNotAModelName.

Since the name is now known when the model is created, I decided that
methods creating a new model can take a name as argument. Since those
methods add the model in the model manager, I believe that it's
elegant, since it allows to avoid having to save the deck a second
time later. Of course, the name is an optional argument, and the
effect is not changed when no name is given.