Do not ask confirmation for language interface
On the first run, a popup window appears asking in which language do
we wish to use Anki.
After language selection, a confirmation was asked.

This commit removes the confirmation as it does not bring much added
value while reducing the first impression (pop-up after the first user
choice is not very encouraging).

As explained on [1], the confirmation dialogs are useful for action
with serious consequences such as actions that can not be undone.
Changing the language is still possible after the confirmation, this
is a rather standard feature and we can assume that a user that
selected the wrong language could figure out how to change the
language back (it is the first option in a standard "Preferences" menu
as many software do).

[1] https://www.nngroup.com/articles/confirmation-dialog/