New model can be edited without full sync
This commit solves a problem that I had many time in the past. When I
create a new model, I usually want to edit it. Clone of existing
models present no interest by themselves. And as soon as I edit it, I
need to do a full sync.

As far as I understand ankiweb (which is sadly closed source), the
full sync is required because ankiweb needs to know that the model
associated to note type on the server did change. But since the model
is new, it has no note type associated to on the server, so there is
no need to do a full sync immediatly. Since the model is new, it also
means there is no risk of the inconsistency with a change made in
another computer/smartphone.

Thus, when a model is created (standard model or cloning) I set it
a value called "ls". This value is equal to col.ls. As long as both
value as equal, it means no sync did occur, and so it can be changed
for free.

Before upload, this value is removed. Indeed, it becomes useless.

Using LS instead of a boolean, for example, has multiple
advantages. Fist, since this value can be removed, it means that
neither ankiweb, nor ankidroid, nor IOS... have to take this into
account (however, I'll quickly port the same thing to ankiweb if and
when it's in Anki). So no change mandatory anywhere else.

But, if for some unfortunate reason, this entry of the dictionnary was
not removed before sync (maybe because of an add-on changing
beforeUpdate, or using a version of Anki without this feature), when
Anki later see the value of LS, it'll knows that a full sync is required.