Conflict Resolution with "--platformtheme"
If you use KDE (or perhaps you just chose the dark theme of Qt design) and you have a dark theme selected, then the interface of Anki will not be displayed correctly. The text in the window will not be readable. Therefore, I use qt5ct for this problem. It allows you to choose a different theme. And for this you need to run the application with the argument "--platformtheme", but if you use it, anki will not start, because it will not understand this argument (although it doesn’t need to do this), or if you use it with one "-", it will see " -p" and it means that the profile is not correct.
Because I just added "--platformtheme" to the arguments parser. This solves it.

P.S 
I ran into this problem on Debian 10.
I apologize because I forgot the default.