_ensureParents consider all cases.
Before this commit, there is a bug done as follow:

Create decks "A::B", "b" and "C".. Now move "b" to "A". You now have a
deck "A::b". Try to move "C" over "A::b". It'll be moved as
"A::B::C". It seems that "A::b" can't get children.

To correct need, I need that "_ensureParents" actually also ensure
that we have the correct case for current deck. I also need to be able
to check the name with correct case without actually creating the
deck; so I added an option to allow getting the case without creating
the deck.

This led to another problem. _ensureParents calls id, which calls
_ensureParents. This extremly quickly led to stack overflow. Since
actually, when _ensureParent calls `id` there is no need to call
_ensureParents since those parents already exists, I added an optional
parameter to state that we know that parents already exists and thus
avoid recomputing them.

This has been tested as follows:
* trying to reproduce the above-mentionned bug and failing
* trying to move and rename deck, and succeeding as much as before