Fix schedv2 _daysSinceCreation
Effective rollover time and calculation of days since creation were
incorrect when localtime has daylight saving time and the interval from
creation to today spans an odd number of transitions between standard
and daylight saving time. The reason for this is that at the transition
between standard and daylight saving time, the day is not 86400 seconds.