Correcting showing errors while importing
An add-on made some importing raise an exception. During the
exception, an other exception was raised, because of this line. It stated:

TypeError: decoding str is not supported

I'm not sure my correction is the good one, because I don't understand
why you want to apply `str` to a string. But if it's an incorrect
correction, at least, this could serve as a bug report