Fix days_elapsed
The definition of days_elapsed is "the number of times the day rolled
over between two timestamps". For example, if start time is a minute
before rollover time on a given day, and end time is rollover time or
later, then one day rolled over "between" start and end time - where
between includes the end time.

All times are given as unix timestamps (seconds past the epoch), but
rollover time, from user perspective, is an hour of the day in local
timezone.

Due to changes in offset from UTC for daylight saving time, not all days
are 86400 seconds long. Therefore, simple division of the seconds
between start and end times does not give correct results in all cases.

Additional (probably excessive) additional tests are provided, including
tests of various times of day that were handled incorrectly with the
previous implementation.

I don't know rust, so the implementation is likely to be less than
ideal, but it is at least a proof of concept.